fix: pass functions as prop instead of redefining in each task
