+++ src/pages/Todolist/components/Task.js
@@ -1,25 +1,9 @@
-import React, { useEffect, useMemo, useState } from 'react';
+import React from 'react';
 
 function Task(props) {
     const { title, description, date_created, priority, id, completed, open, editing, 
-        removeTask, completeTask, openTask, startEditingTask } = props;
-
-    // function to ensure first letter of string is capitalized
-    const capitalize = (string) => string ? string.charAt(0).toUpperCase() + string.slice(1) : "";
-
-    // handling resizing of window for title/description length
-    const [windowWidth, setWindowWidth] = useState(window.innerWidth);
-
-    useEffect(() => {
-        const handleWindowResize = () => setWindowWidth(window.innerWidth);
-        window.addEventListener("resize", handleWindowResize);
-
-        return () => window.removeEventListener("resize", handleWindowResize);
-    })
-
-    // max lengths for task title and description
-    const maxTitleLength = useMemo(() => 11 + windowWidth / 45, [windowWidth]);
-    const maxDescriptionLength = useMemo(() => (windowWidth / 1.2) - (0.000275 * windowWidth ** 2), [windowWidth]);
+        removeTask, completeTask, openTask, startEditingTask,
+        capitalize, maxTitleLength, maxDescriptionLength } = props;
     
     return (
         <div className={`todolist-item priority-${priority}
